Rename and refactor of earlier change
