+++ c/src/core/framing.c
@@ -22,6 +22,7 @@
 #include <stdio.h>
 #include <string.h>
 
+#include "engine-internal.h"
 #include "framing.h"
 
 // TODO: These are near duplicates of code in codec.c - they should be
@@ -105,4 +106,114 @@ size_t pn_write_frame(pn_buffer_t* buffer, pn_frame_t frame)
   } else {
     return 0;
   }
+}
+
+
+static void pn_do_tx_trace(pn_logger_t *logger, uint16_t ch, pn_data_t *args)
+{
+  if (PN_SHOULD_LOG(logger, PN_SUBSYSTEM_AMQP, PN_LEVEL_FRAME) ) {
+    if (pn_data_size(args)==0) {
+      pn_logger_logf(logger, PN_SUBSYSTEM_AMQP, PN_LEVEL_FRAME, "%u -> (EMPTY FRAME)", ch);
+    } else {
+      pni_logger_log_msg_inspect(logger, PN_SUBSYSTEM_AMQP, PN_LEVEL_FRAME, args, "%u -> ", ch);
+    }
+  }
+}
+
+static void pn_do_rx_trace(pn_logger_t *logger, uint16_t ch, pn_data_t *args)
+{
+  if (PN_SHOULD_LOG(logger, PN_SUBSYSTEM_AMQP, PN_LEVEL_FRAME) ) {
+    if (pn_data_size(args)==0) {
+      pn_logger_logf(logger, PN_SUBSYSTEM_AMQP, PN_LEVEL_FRAME, "%u <- (EMPTY FRAME)", ch);
+    } else {
+      pni_logger_log_msg_inspect(logger, PN_SUBSYSTEM_AMQP, PN_LEVEL_FRAME, args, "%u <- ", ch);
+    }
+  }
+}
+
+static void pn_do_trace_payload(pn_logger_t *logger, pn_bytes_t payload)
+{
+  if (PN_SHOULD_LOG(logger, PN_SUBSYSTEM_AMQP, PN_LEVEL_FRAME) ) {
+    char buf[1100];
+    if (payload.size) {
+      size_t l = snprintf(buf, 21, " (%zu) ", payload.size);
+      ssize_t e = pn_quote_data(buf+l, 1024, payload.start, payload.size);
+      if (e == PN_OVERFLOW) {
+        size_t s = strlen(buf);
+        strcpy(buf+s, "... (truncated)");
+      }
+    }
+    pni_logger_log(logger, PN_SUBSYSTEM_AMQP, PN_LEVEL_FRAME, buf);
+  }
+}
+
+static inline void pn_do_raw_trace(pn_logger_t *logger, pn_buffer_t *output, size_t size)
+{
+  PN_LOG_RAW(logger, PN_SUBSYSTEM_IO, PN_LEVEL_RAW, output, size);
+}
+
+static inline void pn_post_frame(pn_buffer_t *output, uint8_t type, uint16_t ch, pn_bytes_t frame_payload)
+{
+  pn_frame_t frame = {
+    .type = type,
+    .channel = ch,
+    .payload = frame_payload.start,
+    .size = frame_payload.size
+  };
+  pn_buffer_ensure(output, AMQP_HEADER_SIZE+frame.ex_size+frame.size);
+  pn_write_frame(output, frame);
+}
+
+int pn_framing_send_amqp(pn_transport_t *transport, uint16_t ch, pn_bytes_t performative)
+{
+  if (!performative.start)
+    return PN_ERR;
+
+  pn_do_tx_trace(&transport->logger, ch, transport->output_args);
+
+  pn_post_frame(transport->output_buffer, AMQP_FRAME_TYPE, ch, performative);
+  pn_do_raw_trace(&transport->logger, transport->output_buffer, AMQP_HEADER_SIZE+performative.size);
+  transport->output_frames_ct += 1;
+  return 0;
+}
+
+int pn_framing_send_amqp_with_payload(pn_transport_t *transport, uint16_t ch, pn_bytes_t performative, pn_bytes_t payload)
+{
+  if (!performative.start)
+    return PN_ERR;
+
+  pn_do_tx_trace(&transport->logger, ch, transport->output_args);
+  pn_do_trace_payload(&transport->logger, payload);
+
+  // For the moment this is a nasty hack - we know that performative is in transport->frame
+  pn_buffer_t *frame = transport->frame;
+  if (pn_buffer_available( frame ) < (payload.size + performative.size)) {
+    // not enough room for payload - resize...
+    // - importantly the performative is still at the front
+    pn_buffer_ensure( frame, payload.size + performative.size );
+  }
+  pn_rwbytes_t wbuf = pn_buffer_memory(frame);
+  wbuf.size = pn_buffer_available(frame);
+  memcpy( wbuf.start + performative.size, payload.start, payload.size);
+
+  pn_post_frame(transport->output_buffer, AMQP_FRAME_TYPE, ch,
+                (pn_bytes_t){.size = performative.size + payload.size, .start = wbuf.start});
+  pn_do_raw_trace(&transport->logger, transport->output_buffer,
+                  AMQP_HEADER_SIZE+performative.size+payload.size);
+  transport->output_frames_ct += 1;
+  return 0;
+}
+
+int pn_framing_send_sasl(pn_transport_t *transport, pn_bytes_t performative)
+{
+  if (!performative.start)
+    return PN_ERR;
+
+  pn_do_tx_trace(&transport->logger, 0, transport->output_args);
+
+  // All SASL frames go on channel 0
+  pn_post_frame(transport->output_buffer, SASL_FRAME_TYPE, 0, performative);
+  pn_do_raw_trace(&transport->logger, transport->output_buffer, AMQP_HEADER_SIZE+performative.size);
+  transport->output_frames_ct += 1;
+  return 0;
 }
